requirements.txt
azure-functions
requests
pandas
pyodbc
azure-storage-blob